requirements.txt
alembic==1.14.0
anyio==3.6.1
apache-airflow==2.10.4
apache-airflow-providers-apache-spark==5.0.0
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==2.1.3
apispec==6.8.1
argcomplete==2.0.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
Babel==2.10.1
blinker==1.9.0
cachelib==0.9.0
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
clickclick==20.10.2
colorama==0.4.4
colorlog==6.9.0
comm==0.2.2
commonmark==0.9.1
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.2.24
croniter==6.0.0
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.0
Flask-Login==0.6.3
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fsspec==2024.12.0
git-filter-repo==2.47.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.3
importlib_metadata==8.5.0
inflection==0.5.1
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.0.3
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy-object-proxy==1.7.1
limits==4.0.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.24.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.5.0
nest-asyncio==1.6.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
ordered-set==4.1.0
packaging==24.2
parso==0.8.4
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
prison==0.2.1
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==5.9.1
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.5
pycparser==2.21
Pygments==2.12.0
PyJWT==2.10.1
pyparsing==2.4.7
pyrsistent==0.18.1
pyspark==3.3.1
python-daemon==3.1.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==12.4.4
rich-argparse==1.6.0
rpds-py==0.22.3
setproctitle==1.3.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.5.3
stack-data==0.6.3
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.2.0
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.6
urllib3==1.26.9
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.14.1
WTForms==2.3.3
zipp==3.21.0
